agent/requirement.txt
praisonaiagents[llm] 
pymongo 
fastmcp